Refactored as per review comment
