+++ hexagonal-architecture/src/main/java/com/baeldung/design/hex/adapter/driving/ControllerForExternalApplication.java
@@ -4,16 +4,17 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.baeldung.design.hex.business.domain.Item;
-import com.baeldung.design.hex.business.service.IOrderService;
-import com.baeldung.design.hex.port.in.IController;
+import com.baeldung.design.hex.port.in.IOrderService;
 
 @RestController
 @RequestMapping("/api/order")
-public class ControllerForExternalApplication implements IController {
+public class ControllerForExternalApplication {
 
     @Autowired
     IOrderService service;
@@ -21,7 +22,7 @@ public class ControllerForExternalApplication implements IController {
     @Value("${props.external.validCallers}")
     List<String> validCallers;
 
-    @Override
+    @PostMapping("/{caller}")
     public String placeOrder(List<Item> items, String caller) {
         String orderId = null;
         if (validCallers.contains(caller)) {
@@ -30,7 +31,7 @@ public class ControllerForExternalApplication implements IController {
         return orderId;
     }
 
-    @Override
+    @GetMapping("/{orderId}/{caller}")
     public List<Item> getOrderedItems(String orderId, String caller) {
         List<Item> items = null;
         if (validCallers.contains(caller)) {
