+++ hexagonal-architecture/src/main/java/com/baeldung/design/hex/business/service/OrderService.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
 import com.baeldung.design.hex.business.domain.Item;
+import com.baeldung.design.hex.port.in.IOrderService;
 import com.baeldung.design.hex.port.out.IMessageSender;
 import com.baeldung.design.hex.port.out.IOrderRepository;
 import com.baeldung.design.hex.port.out.IPrinter;
