+++ hexagonal-architecture/src/main/java/com/baeldung/design/hex/port/in/IOrderService.java
@@ -1,0 +1,11 @@
+package com.baeldung.design.hex.port.in;
+
+import java.util.List;
+
+import com.baeldung.design.hex.business.domain.Item;
+
+public interface IOrderService {
+    public String processOrder(List<Item> items, String caller);
+
+    public List<Item> getOrderedItems(String orderId);
+}
